pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pnn-vit"
version = "0.1.0"
description = "PNN-ViT research prototype for adaptive token routing."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "accelerate>=1.13.0",
  "matplotlib>=3.8.0",
  "PyYAML>=6.0.0",
  "timm>=1.0.24",
  "torch>=2.9.0",
  "torchvision>=0.24.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
